requirements.txt
PyQt5==5.15.9
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
psutil==5.9.6
pytest==7.4.0
pytest-cov==4.1.0
